Address vpn_name issue for systemd
The systemd backend didn't have access to the vpn_name setting that is
passed into __init__.py.  This isn't a super graceful fix, but it should
work until we can come up with something better.  The "name" setting
that gets passed to the systemd backend is essentially the same as the
"vpn_name" setting being passed to __init__.py, except that "vpn_name"
is what is displayed in the bar while "name" is going to be the name
systemd uses to manage the service.