Openvpn: init backends differently
This is another attempt to initialize the systemd backend slightly
differently than the conffile backend.  The idea here is that when
openvpn goes through it's __init__, it will parse what backend is being
loaded and if it's systemd, it will pass the vpn_name into the backend
init.  This then makes vpn_name available to the backend, which is
needed for the systemd backend, but not required for conffile backend.