OpenVPN: fix up unittests
Cleaning up uneeded complexity and hacks in openvpn unittests.  Instead
of relying on the system running the tests to have a specific PID
running (previously, pid 1), I'm just mocking out the call and returning
an arbitrary PID list.  This removes dependencies on the system running
the test while still executing the majority of the function.